requirements.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.31
watchdog==2.1.6
Werkzeug==2.0.2
